+++ src/test/java/com/bebis/BeBiS/item/ItemTestData.java
@@ -3,7 +3,7 @@ package com.bebis.BeBiS.item;
 import com.bebis.BeBiS.integration.blizzard.dto.Item;
 
 public class ItemTestData {
-    public static final int THUNDERFURY_ID = 19019;
+    public static final long THUNDERFURY_ID = 19019;
     public static final String THUNDERFURY_NAME = "Thunderfury, Blessed Blade of the Windseeker";
 
     public static Item thunderfury() {
